Fix CORS pre-flight request handler
I don't know if you already caught this, but I've moved the call to the WriteHeader function to the end of the handleCors function so that it isn't called until after all of the Headers are added to the responseWriter Headers. This makes sure that these headers are added during a pre-flight Options request before the call to WriteHeader which writes the headers into the response.