Ignore content-type parameter during decoder match
I found I had to add a decoder function that matched on
"application/json;charset=UTF-8" because a client had added the charset
parameter to the Content-Type header.

I have changed contentDecoder func in context.go so that it will now
preferentially match against the full Content-Type, but if it finds no
match it will match against the mediaType only.

Whilst I was there I added tests to cover this new functionality and
an additional test to increase the coverage of context.go to 100%

The reason context.go was not at 100% coverage was because the
acceptableMediaTypes function had not been tested for the situation
where the accept header contains an invalid mediaType. In order to
deal with this situation properly I had to add a function to return
whether a mediatype is empty and ignore empty mediatype entries in
the acceptableMediaTypes func. I also added unit tests for this new
error handling functionality.